Chore: Separated partials files from application views in own partials folder
