+++ apps/application/views/partials/application_add_modal_view.py
