+++ apps/application/views/partials/application_edit_modal_view.py
