+++ apps/application/views/partials/application_list_partial_view.py
